修改一些警告

+++ M8Tool/M8Tool/CommonUtil/CommonUtil.m
@@ -118,7 +118,7 @@
         }
     }
     
-    NSString *des = [NSString stringWithFormat:@"逻辑错误，这里应该存满%ld天", lastWeekdays];
+    NSString *des = [NSString stringWithFormat:@"逻辑错误，这里应该存满%ld天", (unsigned long)lastWeekdays];
     BOOL condition = (fifthDaysArr.count == lastWeekdays);
     NSAssert(condition, des);
